Refactoring of Storage - Addresses
