+++ src/main/java/com/iota/iri/service/storage/Storage.java
@@ -21,7 +21,6 @@ public class Storage extends AbstractStorage {
 	
 	private static final Logger log = LoggerFactory.getLogger(Storage.class);
 
-    private static final String ADDRESSES_FILE_NAME = "addresses.iri";
     private static final String TAGS_FILE_NAME = "tags.iri";
     private static final String APPROVERS_FILE_NAME = "approvers.iri";
     private static final String SCRATCHPAD_FILE_NAME = "scratchpad.iri";
@@ -29,11 +28,7 @@ public class Storage extends AbstractStorage {
     public static final byte[][] approvedTransactionsToStore = new byte[2][];
     
     public static int numberOfApprovedTransactionsToStore;
-
-    private FileChannel addressesChannel;
-    private final ByteBuffer[] addressesChunks = new ByteBuffer[MAX_NUMBER_OF_CHUNKS];
-    private volatile long addressesNextPointer = SUPER_GROUPS_SIZE;
-
+    
     private FileChannel tagsChannel;
     private final ByteBuffer[] tagsChunks = new ByteBuffer[MAX_NUMBER_OF_CHUNKS];
     private volatile long tagsNextPointer = SUPER_GROUPS_SIZE;
@@ -55,43 +50,15 @@ public class Storage extends AbstractStorage {
     
     private StorageTransactions storageTransactionInstance = StorageTransactions.instance();
     private StorageBundle storageBundleInstance = StorageBundle.instance();
-
+    private StorageAddresses storageAddressesInstance = StorageAddresses.instance();
+    
     @Override
     public synchronized void init() throws IOException {
     	
     	storageTransactionInstance.init();
     	storageBundleInstance.init();
+    	storageAddressesInstance.init();
     	
-        addressesChannel = FileChannel.open(Paths.get(ADDRESSES_FILE_NAME), StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
-        addressesChunks[0] = addressesChannel.map(FileChannel.MapMode.READ_WRITE, 0, SUPER_GROUPS_SIZE);
-        final long addressesChannelSize = addressesChannel.size();
-        while (true) {
-
-            if ((addressesNextPointer & (CHUNK_SIZE - 1)) == 0) {
-                addressesChunks[(int)(addressesNextPointer >> 27)] = addressesChannel.map(FileChannel.MapMode.READ_WRITE, addressesNextPointer, CHUNK_SIZE);
-            }
-
-            if (addressesChannelSize - addressesNextPointer > CHUNK_SIZE) {
-                addressesNextPointer += CHUNK_SIZE;
-            } else {
-
-                addressesChunks[(int) (addressesNextPointer >> 27)].get(mainBuffer);
-                boolean empty = true;
-                for (final int value : mainBuffer) {
-
-                    if (value != 0) {
-                        empty = false;
-                        break;
-                    }
-                }
-                if (empty) {
-                    break;
-                }
-
-                addressesNextPointer += CELL_SIZE;
-            }
-        }
-
         tagsChannel = FileChannel.open(Paths.get(TAGS_FILE_NAME), StandardOpenOption.CREATE, StandardOpenOption.READ, StandardOpenOption.WRITE);
         tagsChunks[0] = tagsChannel.map(FileChannel.MapMode.READ_WRITE, 0, SUPER_GROUPS_SIZE);
         final long tagsChannelSize = tagsChannel.size();
@@ -169,12 +136,7 @@ public class Storage extends AbstractStorage {
 
         	storageTransactionInstance.shutdown();
         	storageBundleInstance.shutdown();
-        	
-
-            for (int i = 0; i < MAX_NUMBER_OF_CHUNKS && addressesChunks[i] != null; i++) {
-            	log.info("Flushing addresses chunk #" + i);
-                flush(addressesChunks[i]);
-            }
+        	storageAddressesInstance.shutdown();
 
             for (int i = 0; i < MAX_NUMBER_OF_CHUNKS && tagsChunks[i] != null; i++) {
             	log.info("Flushing tags chunk #" + i);
@@ -189,7 +151,6 @@ public class Storage extends AbstractStorage {
             log.info("DB successfully flushed");
 
             try {
-                addressesChannel.close();
                 tagsChannel.close();
                 approversChannel.close();
 
@@ -285,71 +246,6 @@ public class Storage extends AbstractStorage {
         analyzedTransactionsFlagsCopy.position(0);
         analyzedTransactionsFlags.position(0);
         analyzedTransactionsFlags.put(analyzedTransactionsFlagsCopy);
-    }
-
-    public synchronized long addressPointer(final byte[] hash) {
-
-        long pointer = ((hash[0] + 128) + ((hash[1] + 128) << 8)) << 11;
-        for (int depth = 2; depth < Transaction.ADDRESS_SIZE; depth++) {
-
-            ((ByteBuffer)addressesChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-
-            if (mainBuffer[Transaction.TYPE_OFFSET] == GROUP) {
-
-                if ((pointer = value(mainBuffer, (hash[depth] + 128) << 3)) == 0) {
-                    return 0;
-                }
-            } else {
-
-                for (; depth < Transaction.ADDRESS_SIZE; depth++) {
-
-                    if (mainBuffer[Transaction.HASH_OFFSET + depth] != hash[depth]) {
-                        return 0;
-                    }
-                }
-                return pointer;
-            }
-        }
-
-        throw new IllegalStateException("Corrupted storage");
-    }
-
-    public synchronized List<Long> addressTransactions(final long pointer) {
-
-        final List<Long> addressTransactions = new LinkedList<>();
-
-        if (pointer != 0) {
-
-            ((ByteBuffer) addressesChunks[(int) (pointer >> 27)].position((int) (pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-            int offset = ZEROTH_POINTER_OFFSET - Long.BYTES;
-            while (true) {
-
-                while ((offset += Long.BYTES) < CELL_SIZE - Long.BYTES) {
-
-                    final long transactionPointer = value(mainBuffer, offset);
-                    if (transactionPointer == 0) {
-                        break;
-                    } else {
-                        addressTransactions.add(transactionPointer);
-                    }
-                }
-                if (offset == CELL_SIZE - Long.BYTES) {
-
-                    final long nextCellPointer = value(mainBuffer, offset);
-                    if (nextCellPointer == 0) {
-                        break;
-                    } else {
-                        ((ByteBuffer) addressesChunks[(int) (nextCellPointer >> 27)].position((int) (nextCellPointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                        offset = -Long.BYTES;
-                    }
-
-                } else {
-                    break;
-                }
-            }
-        }
-
-        return addressTransactions;
     }
 
     public synchronized long tagPointer(final byte[] hash) {
@@ -490,28 +386,13 @@ public class Storage extends AbstractStorage {
         return approveeTransactions;
     }
 
-    private void appendToAddresses() {
-
-        ((ByteBuffer)addressesChunks[(int)(addressesNextPointer >> 27)].position((int)(addressesNextPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-        if (((addressesNextPointer += CELL_SIZE) & (CHUNK_SIZE - 1)) == 0) {
-
-            try {
-                addressesChunks[(int)(addressesNextPointer >> 27)] = addressesChannel.map(FileChannel.MapMode.READ_WRITE, addressesNextPointer, CHUNK_SIZE);
-            } catch (final IOException e) {
-            	log.error("Caught exception on appendToAddresses:", e);
-            }
-        }
-    }
-
     private void appendToTags() {
 
         ((ByteBuffer) tagsChunks[(int)(tagsNextPointer >> 27)].position((int)(tagsNextPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
         if (((tagsNextPointer += CELL_SIZE) & (CHUNK_SIZE - 1)) == 0) {
 
             try {
-
                 tagsChunks[(int)(tagsNextPointer >> 27)] = tagsChannel.map(FileChannel.MapMode.READ_WRITE, tagsNextPointer, CHUNK_SIZE);
-
             } catch (final IOException e) {
             	log.error("Caught exception on appendToTags:", e);
             }
@@ -541,111 +422,7 @@ public class Storage extends AbstractStorage {
 
         StorageBundle.instance().updateBundle(transactionPointer, transaction);
 
-        {
-            long pointer = ((transaction.address[0] + 128) + ((transaction.address[1] + 128) << 8)) << 11, prevPointer = 0;
-            for (int depth = 2; depth < Transaction.ADDRESS_SIZE; depth++) {
-
-                ((ByteBuffer)addressesChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-
-                if (mainBuffer[Transaction.TYPE_OFFSET] == GROUP) {
-
-                    prevPointer = pointer;
-                    if ((pointer = value(mainBuffer, (transaction.address[depth] + 128) << 3)) == 0) {
-
-                        setValue(mainBuffer, (transaction.address[depth] + 128) << 3, addressesNextPointer);
-                        ((ByteBuffer)addressesChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                        System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                        mainBuffer[Transaction.TYPE_OFFSET] = FILLED_SLOT;
-                        System.arraycopy(transaction.address, 0, mainBuffer, 8, Transaction.ADDRESS_SIZE);
-                        setValue(mainBuffer, ZEROTH_POINTER_OFFSET, transactionPointer);
-                        appendToAddresses();
-
-                        break;
-                    }
-
-                } else {
-
-                    boolean sameAddress = true;
-
-                    for (int i = depth; i < Transaction.ADDRESS_SIZE; i++) {
-
-                        if (mainBuffer[Transaction.HASH_OFFSET + i] != transaction.address[i]) {
-
-                            final int differentHashByte = mainBuffer[Transaction.HASH_OFFSET + i];
-
-                            ((ByteBuffer)addressesChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                            setValue(mainBuffer, (transaction.address[depth - 1] + 128) << 3, addressesNextPointer);
-                            ((ByteBuffer)addressesChunks[(int)(prevPointer >> 27)].position((int)(prevPointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                            for (int j = depth; j < i; j++) {
-
-                                System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                setValue(mainBuffer, (transaction.address[j] + 128) << 3, addressesNextPointer + CELL_SIZE);
-                                appendToAddresses();
-                            }
-
-                            System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                            setValue(mainBuffer, (differentHashByte + 128) << 3, pointer);
-                            setValue(mainBuffer, (transaction.address[i] + 128) << 3, addressesNextPointer + CELL_SIZE);
-                            appendToAddresses();
-
-                            System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                            mainBuffer[Transaction.TYPE_OFFSET] = FILLED_SLOT;
-                            System.arraycopy(transaction.address, 0, mainBuffer, 8, Transaction.ADDRESS_SIZE);
-                            setValue(mainBuffer, ZEROTH_POINTER_OFFSET, transactionPointer);
-                            appendToAddresses();
-
-                            sameAddress = false;
-
-                            break;
-                        }
-                    }
-
-                    if (sameAddress) {
-
-                        int offset = ZEROTH_POINTER_OFFSET;
-                        while (true) {
-
-                            while ((offset += Long.BYTES) < CELL_SIZE - Long.BYTES && value(mainBuffer, offset) != 0) {
-
-                                // Do nothing
-                            }
-                            if (offset == CELL_SIZE - Long.BYTES) {
-
-                                final long nextCellPointer = value(mainBuffer, offset);
-                                if (nextCellPointer == 0) {
-
-                                    setValue(mainBuffer, offset, addressesNextPointer);
-                                    ((ByteBuffer)addressesChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                                    System.arraycopy(ZEROED_BUFFER, 0, mainBuffer, 0, CELL_SIZE);
-                                    setValue(mainBuffer, 0, transactionPointer);
-                                    appendToAddresses();
-
-                                    break;
-
-                                } else {
-
-                                    pointer = nextCellPointer;
-                                    ((ByteBuffer)addressesChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).get(mainBuffer);
-                                    offset = -Long.BYTES;
-                                }
-
-                            } else {
-
-                                setValue(mainBuffer, offset, transactionPointer);
-                                ((ByteBuffer)addressesChunks[(int)(pointer >> 27)].position((int)(pointer & (CHUNK_SIZE - 1)))).put(mainBuffer);
-
-                                break;
-                            }
-                        }
-                    }
-
-                    break;
-                }
-            }
-        }
+        StorageAddresses.instance().updateAddresses(transactionPointer, transaction);
 
         for (int i = 0; i < Transaction.TAG_SIZE; i++) {
 
@@ -765,7 +542,7 @@ public class Storage extends AbstractStorage {
         }
     }
 
-    private void updateApprover(final byte[] hash, final long transactionPointer) {
+	private void updateApprover(final byte[] hash, final long transactionPointer) {
 
         long pointer = ((hash[0] + 128) + ((hash[1] + 128) << 8)) << 11, prevPointer = 0;
         for (int depth = 2; depth < Transaction.HASH_SIZE; depth++) {
